requirements.txt
# Core packages
streamlit>=1.28.0
openai>=1.3.0
python-dotenv>=1.0.0

# Data processing
pandas>=2.0.0
numpy>=1.24.0
requests>=2.31.0

# Database clients
redis>=5.0.0
psycopg2-binary>=2.9.0
elasticsearch>=8.10.0
qdrant-client>=1.6.0

# Optional performance packages
watchdog>=3.0.0
